api/pyproject.toml
[project]
name = "dirt-dossier-api"
version = "0.1.0"
description = "dirt-dossier backend: Strava ingest, trail matching, FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "geoalchemy2>=0.15",
    "shapely>=2.0",
    "asyncpg>=0.29",
    "httpx>=0.27",
    "fastkml>=1.0",
    "lxml>=5.0",
    "python-dotenv>=1.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.13.1",
    "osmium>=4.3.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dirt_dossier"]
